Added "within metro area" flag to route segments.
Routes inside the metro area are now given priority. Routes outside the metro area may only be selected if all routes inside the metro area were selected first.

Also, updated the routes in the DB to match the new flag. Added the rest of the routes in LA and Orange counties and split some other routes at the county line.